refactors and fixes some tests
